[Bronze III] Title: 중앙 이동 알고리즘, Time: 104 ms, Memory: 14116 KB -BaekjoonHub

+++ 백준/Bronze/2903. 중앙 이동 알고리즘/중앙 이동 알고리즘.java
@@ -4,7 +4,7 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
-        double a = Double.parseDouble(br.readLine());
+        int a = Integer.parseInt(br.readLine());
         System.out.println((int) Math.pow(Math.pow(2,a) + 1, 2));
     }
 }
